DMSS - Database management system for security
Welcome Message from Zakey

This module will introduce you to the exciting world of Databases and Database Security. As this is a fundamental topic in IT it is very important that you take this module seriously and study hard. I have taught this module (or similar modules) for 5-6 years and it has been really a fun experience, since this is one of my favorite topics as well.